pyproject.toml
[project]
name = "cricket-nepal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.41",
]
